[Review] 디폴트 생성자 추가

+++ Review/Review/Main.cpp
@@ -59,7 +59,7 @@ int main()
 		// ·»´õ¸µ
 		shader.use();
 
-		float timeValue = glfwGetTime();
+		float timeValue = (float)glfwGetTime();
 		float greenValue = sin(timeValue) / 2.0f * 0.5f;
 		shader.setUniform4f("ourColor", 0.0f, greenValue, 0.0f, 1.0f);
